--- iosHyperskillApp/iosHyperskillApp/Assets.xcassets/StepQuiz/ProblemsLimitReachedModal/problems-limit-reached-modal-cube.imageset/Contents.json
@@ -1,25 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "problems-limit-reached-modal-cube-light.pdf",
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "problems-limit-reached-modal-cube-dark.pdf",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "preserves-vector-representation" : true
-  }
-}
