--- iosHyperskillApp/iosHyperskillApp/Sources/Modules/StepQuiz/Views/Modals/ProblemsLimitReachedModalViewControllerDelegate.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-protocol ProblemsLimitReachedModalViewControllerDelegate: AnyObject {
-    func problemsLimitReachedModalViewControllerDidTapGoToHomescreenButton(
-        _ viewController: ProblemsLimitReachedModalViewController
-    )
-
-    func problemsLimitReachedModalViewControllerDidAppear(_ viewController: ProblemsLimitReachedModalViewController)
-
-    func problemsLimitReachedModalViewControllerDidDisappear(_ viewController: ProblemsLimitReachedModalViewController)
-}
