+++ shared/src/commonMain/kotlin/org/hyperskill/app/problems_limit/presentation/ProblemsLimitFeature.kt
@@ -58,6 +58,6 @@ object ProblemsLimitFeature {
 
         data class LaunchTimer(val updateIn: Duration) : Action
 
-        sealed interface ViewAction
+        sealed interface ViewAction : Action
     }
 }
